IGAPP-622: improve capabilities setup & add run 'new e2e tests' in ci
