+++ .circleci/src/workflows/commit.yml
@@ -8,6 +8,14 @@ jobs:
   - check_web:
       context:
         - mattermost
+  - e2e_web:
+      name: e2e_web
+      build_config_name: integreat-test-cms
+      context:
+        - mattermost
+        - requires:
+            - bump_version
+            - check_web
   - build_web:
       name: build_ig_test_cms_web
       build_config_name: integreat-test-cms
