+++ e2e-tests/web/browserstack/wdio-browserstack-chrome.conf.ts
@@ -1,5 +1,5 @@
 import { config as defaultConfig } from './wdio-browserstack.conf'
 
 const config = { ...defaultConfig }
-config.capabilities = config.capabilities.filter(cap => cap.browserName?.toLowerCase() === 'chrome')
+config.capabilities = config.capabilities.filter(cap => cap.browserName.toLowerCase() === 'chrome')
 export { config }
