+++ e2e-tests/web/browserstack/wdio-browserstack-ie.conf.ts
@@ -1,6 +1,6 @@
 import { config as defaultConfig } from './wdio-browserstack.conf'
 
 const config = { ...defaultConfig }
-config.capabilities = config.capabilities.filter(cap => cap.browserName?.toLowerCase() === 'ie')
+config.capabilities = config.capabilities.filter(cap => cap.browserName.toLowerCase() === 'ie')
 
 export { config }
