+++ native/package.json
@@ -28,7 +28,7 @@
     "test:ci": "yarn run test:coverage --ci",
     "test:watch": "yarn run test --watchAll",
     "test:update": "yarn run test -u",
-    "test:e2e": "cross-env BUILD_CONFIG_NAME='integreat-e2e' jest --config jest.e2e.config.ts --maxWorkers=4",
+    "test:e2e": "yarn workspace e2e test:native:browserstack",
     "lint": "yarn run eslint",
     "lint:ci": "yarn run eslint --quiet --format junit -o reports/lint/junit-lint.xml",
     "eslint": "eslint .",
